+++ components/Profile.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { Avatar, AvatarFallback, AvatarImage } from "../avatar";
+import { Avatar, AvatarFallback, AvatarImage } from "./avatar";
 
 interface ProfileProps {
   imageUrl: string;
